More refactoring
Split code into separate files and use the fact that we now have a compile time list of log topics with constant ids.
