+++ arangod/RestServer/DaemonFeature.cpp
@@ -270,7 +270,7 @@ int DaemonFeature::forkProcess() {
   TRI_ASSERT(pid == 0);  // we are in the child
 
   // child
-  LogAppender::allowStdLogging(false);
+  Logger::allowStdLogging(false);
   Logger::clearCachedPid();
 
   // change the file mode mask
