+++ arangod/Scheduler/SchedulerFeature.cpp
@@ -480,7 +480,7 @@ extern "C" void c_hangup_handler(int signal, siginfo_t* info, void*) {
           LOG_TOPIC("33eae", INFO, arangodb::Logger::FIXME)
               << "hangup received, about to reopen logfile (sender pid "
               << processIdRequesting << ")";
-          LogAppender::reopen();
+          Logger::reopen();
           LOG_TOPIC("23db2", INFO, arangodb::Logger::FIXME)
               << "hangup received, reopened logfile";
         } catch (...) {
