+++ tests/main.cpp
@@ -127,7 +127,7 @@ int main(int argc, char* argv[]) {
       arangodb::LogTimeFormats::TimeFormat::UTCDateStringMicros);
   arangodb::Logger::setShowThreadIdentifier(true);
   arangodb::Logger::initialize(false, 10000);
-  arangodb::LogAppender::addAppender(arangodb::Logger::defaultLogGroup(), "-");
+  arangodb::Logger::addAppender(arangodb::Logger::defaultLogGroup(), "-");
 
   sc.prepare();
